+++ NEWS.md
@@ -1,3 +1,4 @@
 ## 2021-XX-XX v2.4.0-SNAPSHOT
 * [KAFKAWRAP-2](https://issues.folio.org/browse/KAFKAWRAP-2) Take folio-kafka-wrapper lib out of mod-pubsub repository
 * [KAFKAWRAP-10](https://issues.folio.org/browse/KAFKAWRAP-10) Provide property to set compression type for producer configuration
+* [KAFKAWRAP-11](https://issues.folio.org/browse/KAFKAWRAP-11) Remove duplicate kafka headers